Restructure highlight/ folder
