+++ src/highlights/toggle-highlight/index.ts
@@ -1,4 +1,4 @@
-import { Refactoring } from "../types";
+import { Refactoring } from "../../types";
 import { toggleHighlight } from "./toggle-highlight";
 
 const config: Refactoring = {
