+++ src/highlights/toggle-highlight/toggle-highlight.ts
@@ -1,6 +1,5 @@
-import { Editor } from "../editor/editor";
-import { Selection } from "../editor/selection";
-import * as t from "../ast";
+import { Editor, Selection } from "../../editor";
+import * as t from "../../ast";
 
 export async function toggleHighlight(editor: Editor): Promise<void> {
   const { code, selection } = editor;
